requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
ttkbootstrap
python-dotenv
openai
